problems pulling

+++ src/hhu/propra_2013/gruppe_13/EnemyMelee.java
@@ -23,7 +23,7 @@ public class EnemyMelee extends Enemy{
 	private long 	regenerate;
 	
 	// Standard constructor, build enemies according to input tyoe
-	EnemyMelee(double inx, double iny,int inWidth, int inHeight, int inType, int inStage){
+	EnemyMelee(double inx, double iny,double inWidth, double inHeight, int inType, int inStage){
 		x 			= inx;
 		y 			= iny;
